VB-4368 - Move event publishing outside of transaction
Create a new service which handles storing and updating visit
data, this allows us to remove the transaction at the top level
function where the events are called to be processed, now outside
of any transaction
